Simplify Providers' namespaces
