+++ docs/en/developer/comparator.txt
@@ -1,7 +1,7 @@
 Comparator
 ==========
 
-.. php:namespace:: Baleen\Cli\Container
+.. php:namespace:: Baleen\Cli\Provider
 
 Service Name: :php:const:`Services::COMPARATOR`
