+++ docs/en/developer/configuration.txt
@@ -1,7 +1,7 @@
 Configuration
 =============
 
-.. php:namespace:: Baleen\Cli\Container
+.. php:namespace:: Baleen\Cli\Provider
 
 Service Name: :php:const:`Services::CONFIG`
