--- docs/en/reference/Baleen/Cli/Container/ServiceProvider/HelperSetProvider.txt
@@ -1,15 +1,0 @@
-----------------------------------------------------------
-Baleen\\Cli\\Container\\ServiceProvider\\HelperSetProvider
-----------------------------------------------------------
-
-.. php:namespace: Baleen\\Cli\\Container\\ServiceProvider
-
-.. php:class:: HelperSetProvider
-
-    Class HelperSetProvider.
-
-    .. php:attr:: provides
-
-        protected
-
-    .. php:method:: register()
