--- docs/en/reference/Baleen/Cli/Container/ServiceProvider/RepositoryProvider.txt
@@ -1,15 +1,0 @@
------------------------------------------------------------
-Baleen\\Cli\\Container\\ServiceProvider\\RepositoryProvider
------------------------------------------------------------
-
-.. php:namespace: Baleen\\Cli\\Container\\ServiceProvider
-
-.. php:class:: RepositoryProvider
-
-    Class RepositoryProvider.
-
-    .. php:attr:: provides
-
-        protected
-
-    .. php:method:: register()
