--- docs/en/reference/Baleen/Cli/Container/ServiceProvider/StorageProvider.txt
@@ -1,15 +1,0 @@
---------------------------------------------------------
-Baleen\\Cli\\Container\\ServiceProvider\\StorageProvider
---------------------------------------------------------
-
-.. php:namespace: Baleen\\Cli\\Container\\ServiceProvider
-
-.. php:class:: StorageProvider
-
-    Class StorageProvider.
-
-    .. php:attr:: provides
-
-        protected
-
-    .. php:method:: register()
