--- docs/en/reference/Baleen/Cli/Container/ServiceProvider/TimelineProvider.txt
@@ -1,15 +1,0 @@
----------------------------------------------------------
-Baleen\\Cli\\Container\\ServiceProvider\\TimelineProvider
----------------------------------------------------------
-
-.. php:namespace: Baleen\\Cli\\Container\\ServiceProvider
-
-.. php:class:: TimelineProvider
-
-    Class TimelineProvider.
-
-    .. php:attr:: provides
-
-        protected
-
-    .. php:method:: register()
