--- docs/en/reference/Baleen/Cli/Container/ServiceProvider/index.txt
@@ -1,15 +1,0 @@
-"""""""""""""""""""""""""""""""""""""""
-Baleen\\Cli\\Container\\ServiceProvider
-"""""""""""""""""""""""""""""""""""""""
-
-.. php:namespace: Baleen\\Cli\\Container\\ServiceProvider
-
-.. toctree::
-
-   ApplicationProvider
-   CommandsProvider
-   ConfigProvider
-   HelperSetProvider
-   RepositoryProvider
-   StorageProvider
-   TimelineProvider
