--- docs/en/reference/Baleen/Cli/Container/index.txt
@@ -1,10 +1,0 @@
-''''''''''''''''''''''
-Baleen\\Cli\\Container
-''''''''''''''''''''''
-
-.. php:namespace: Baleen\\Cli\\Container
-
-.. toctree::
-
-   ServiceProvider/index
-   Services
