--- docs/en/reference/Baleen/Cli/Helper/index.txt
@@ -1,9 +1,0 @@
-'''''''''''''''''''
-Baleen\\Cli\\Helper
-'''''''''''''''''''
-
-.. php:namespace: Baleen\\Cli\\Helper
-
-.. toctree::
-
-   ConfigHelper
