+++ docs/en/reference/Baleen/Cli/Provider/RepositoryProvider.txt
@@ -1,0 +1,15 @@
+-----------------------------------------
+Baleen\\Cli\\Provider\\RepositoryProvider
+-----------------------------------------
+
+.. php:namespace: Baleen\\Cli\\Provider
+
+.. php:class:: RepositoryProvider
+
+    Class RepositoryProvider.
+
+    .. php:attr:: provides
+
+        protected
+
+    .. php:method:: register()
