+++ docs/en/reference/Baleen/Cli/index.txt
@@ -10,6 +10,5 @@ Baleen\\Cli
    BaseCommand
    CommandBus/index
    Config/index
-   Container/index
    Exception/index
-   Helper/index
+   Provider/index
